frontend refactor: Reformated CreateEventPage code for readability by moving DatePicker elements to ComplexComponents
